+++ src/gdr/gdr.hpp
@@ -14,8 +14,8 @@ cocos2d::CCPoint dataFromString(std::string dataString);
 
 std::vector<std::string> splitByChar(std::string str, char splitChar);
 
-const std::string geobotVersion = "1.1";
-constexpr bool geobotDisableBuildExpiryLock = true;
+const std::string geobotVersion = "1.1.1-dev";
+constexpr bool geobotDisableBuildExpiryLock = false;
 
 namespace gdr {